Add '-HyperVGeneration $Disk.HyperVGeneration' in New-AzDiskConfig
Add '-HyperVGeneration $Disk.HyperVGeneration' in New-AzDiskConfig. Otherwise might get error when swap the OS disk if the HyperVisor property on the OSDisk and VM do not match:

Update-AzVM : Swapping OS Disk is not supported since HyperVisor property on the OSDisk and VM do not match. Please
make sure the HyperVisor property on the OSDisk is HyperVisor Generation 'V2'